# agnesui.openapi.yml
# created with the help of https://editor.swagger.io/

openapi: 3.0.1
info:
  title: Agnes UI
  version: 1.0.0
servers:
  - url: /api
paths:
  /release:
    post:
      tags:
        - release
      summary: Add a new release
      operationId: add
      requestBody:
        description: The release to be created
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Release'
        required: true
      responses:
        201:
          description: Created
          content: {}


  /release/all:
    get:
      tags:
        - release
      summary: Gets all releases
      operationId: getAll
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Release'

  /instance/all:
    get:
      tags:
        - instance
      summary: gets all instances
      operationId: getAll
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Instance'

  /deloy:
    post:
      tags:
        - deploy
      summary: Deploy to environments
      operationId: deploy
      requestBody:
        description: The deployment to start
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Deployment'
        required: true
      responses:
        201:
          description: Created
          content: {}

  /deloy/dryrun:
    post:
      tags:
        - deploy
      summary: Check which instances the deploy would affect
      operationId: deployDryRun
      requestBody:
        description: The deployment to dry run
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Deployment'
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Instance'

  /rollback:
    post:
      tags:
        - rollback
      summary: Rollback an environment to a previous stage
      operationId: rollback
      requestBody:
        description: The Rollback to execute
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Rollback'
        required: true
      responses:
        201:
          description: Created
          content: {}

  /rollback/dryrun:
    post:
      tags:
        - rollback
      summary: Check which instances the Rollback would affect
      operationId: rollbackDryRun
      requestBody:
        description: The rollback to dry run
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Rollback'
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PendingReleaseInstance'

  /copyShared:
    post:
      tags:
        - copyShared
      summary: Copy the shared data from source to target (replicate source on target).
      operationId: copyShared
      requestBody:
        description: The copy shared action
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CopyShared'
        required: true
      responses:
        201:
          description: Created
          content: {}

  /copyShared/dryrun:
    post:
      tags:
        - copyShared
      summary: Check which instances the copy shared action would affect
      operationId: copySharedDryRun
      requestBody:
        description: The copy shared action to dry run
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CopyShared'
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AcrossInstancesAction'

components:
  schemas:
    Release:
      required:
        - commitish
        - description
        - name
      type: object
      properties:
        commitish:
          type: string
        description:
          type: string
        name:
          type: string
    Deployment:
      required:
        - target
        - release
      type: object
      properties:
        target:
          type: string
        release:
          type: string
    Rollback:
      required:
        - target
      type: object
      properties:
        target:
          type: string
        rollbackFrom:
          type: string
        rollbackTo:
          type: string
    CopyShared:
      required:
        - source
        - target
      type: object
      properties:
        target:
          type: string
        source:
          type: string
    AcrossInstancesAction:
      required:
        - source
        - target
      type: object
      properties:
        source:
          $ref: '#/components/schemas/Instance'
        target:
          $ref: '#/components/schemas/Instance'
    PendingReleaseInstance:
      required:
        - pendingReleaseName
        - instance
      type:  object
      properties:
        pendingReleaseName:
          type: string
        instance:
          $ref: '#/components/schemas/Instance'
    Instance:
      required:
        - server
        - environment
        - stage
      type: object
      properties:
        server:
          type: string
        environment:
          type: string
        stage:
          type: string
        currentReleaseName:
          type: string
    File:
      required:
        - path
        - content
      type: object
      properties:
        path:
          type: string
        content:
          type: string
    Installation:
      type: object
      properties:
        release:
          $ref: '#/components/schemas/Release'