# src/Model/CopyShared.php
<?php
/**
 * CopyShared
 *
 * PHP version 5
 *
 * @category Class
 * @package  App\Model
 * @author   OpenAPI Generator team
 * @link     https://github.com/openapitools/openapi-generator
 */

/**
 * Agnes UI
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 * Generated by: https://github.com/openapitools/openapi-generator.git
 *
 */

/**
 * NOTE: This class is auto generated by the openapi generator program.
 * https://github.com/openapitools/openapi-generator
 * Do not edit the class manually.
 */

namespace App\Model;

use Symfony\Component\Validator\Constraints as Assert;
use JMS\Serializer\Annotation\Type;
use JMS\Serializer\Annotation\SerializedName;

/**
 * Class representing the CopyShared model.
 *
 * @package App\Model
 * @author  OpenAPI Generator team
 */
class CopyShared 
{
        /**
     * @var string
     * @SerializedName("target")
     * @Assert\NotNull()
     * @Assert\Type("string")
     * @Type("string")
     */
    protected $target;

    /**
     * @var string
     * @SerializedName("source")
     * @Assert\NotNull()
     * @Assert\Type("string")
     * @Type("string")
     */
    protected $source;

    /**
     * Constructor
     * @param mixed[] $data Associated array of property values initializing the model
     */
    public function __construct(array $data = null)
    {
        $this->target = isset($data['target']) ? $data['target'] : null;
        $this->source = isset($data['source']) ? $data['source'] : null;
    }

    /**
     * Gets target.
     *
     * @return string
     */
    public function getTarget()
    {
        return $this->target;
    }

    /**
     * Sets target.
     *
     * @param string $target
     *
     * @return $this
     */
    public function setTarget($target)
    {
        $this->target = $target;

        return $this;
    }

    /**
     * Gets source.
     *
     * @return string
     */
    public function getSource()
    {
        return $this->source;
    }

    /**
     * Sets source.
     *
     * @param string $source
     *
     * @return $this
     */
    public function setSource($source)
    {
        $this->source = $source;

        return $this;
    }
}