# src/Model/PendingReleaseInstance.php
<?php
/**
 * PendingReleaseInstance
 *
 * PHP version 5
 *
 * @category Class
 * @package  App\Model
 * @author   OpenAPI Generator team
 * @link     https://github.com/openapitools/openapi-generator
 */

/**
 * Agnes UI
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 * Generated by: https://github.com/openapitools/openapi-generator.git
 *
 */

/**
 * NOTE: This class is auto generated by the openapi generator program.
 * https://github.com/openapitools/openapi-generator
 * Do not edit the class manually.
 */

namespace App\Model;

use Symfony\Component\Validator\Constraints as Assert;
use JMS\Serializer\Annotation\Type;
use JMS\Serializer\Annotation\SerializedName;

/**
 * Class representing the PendingReleaseInstance model.
 *
 * @package App\Model
 * @author  OpenAPI Generator team
 */
class PendingReleaseInstance 
{
        /**
     * @var string
     * @SerializedName("pendingReleaseName")
     * @Assert\NotNull()
     * @Assert\Type("string")
     * @Type("string")
     */
    protected $pendingReleaseName;

    /**
     * @var App\Model\Instance
     * @SerializedName("instance")
     * @Assert\NotNull()
     * @Assert\Type("App\Model\Instance")
     * @Type("App\Model\Instance")
     */
    protected $instance;

    /**
     * Constructor
     * @param mixed[] $data Associated array of property values initializing the model
     */
    public function __construct(array $data = null)
    {
        $this->pendingReleaseName = isset($data['pendingReleaseName']) ? $data['pendingReleaseName'] : null;
        $this->instance = isset($data['instance']) ? $data['instance'] : null;
    }

    /**
     * Gets pendingReleaseName.
     *
     * @return string
     */
    public function getPendingReleaseName()
    {
        return $this->pendingReleaseName;
    }

    /**
     * Sets pendingReleaseName.
     *
     * @param string $pendingReleaseName
     *
     * @return $this
     */
    public function setPendingReleaseName($pendingReleaseName)
    {
        $this->pendingReleaseName = $pendingReleaseName;

        return $this;
    }

    /**
     * Gets instance.
     *
     * @return App\Model\Instance
     */
    public function getInstance()
    {
        return $this->instance;
    }

    /**
     * Sets instance.
     *
     * @param App\Model\Instance $instance
     *
     * @return $this
     */
    public function setInstance(Instance $instance)
    {
        $this->instance = $instance;

        return $this;
    }
}