# src/Model/Rollback.php
<?php
/**
 * Rollback
 *
 * PHP version 5
 *
 * @category Class
 * @package  App\Model
 * @author   OpenAPI Generator team
 * @link     https://github.com/openapitools/openapi-generator
 */

/**
 * Agnes UI
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 * Generated by: https://github.com/openapitools/openapi-generator.git
 *
 */

/**
 * NOTE: This class is auto generated by the openapi generator program.
 * https://github.com/openapitools/openapi-generator
 * Do not edit the class manually.
 */

namespace App\Model;

use Symfony\Component\Validator\Constraints as Assert;
use JMS\Serializer\Annotation\Type;
use JMS\Serializer\Annotation\SerializedName;

/**
 * Class representing the Rollback model.
 *
 * @package App\Model
 * @author  OpenAPI Generator team
 */
class Rollback 
{
        /**
     * @var string
     * @SerializedName("target")
     * @Assert\NotNull()
     * @Assert\Type("string")
     * @Type("string")
     */
    protected $target;

    /**
     * @var string|null
     * @SerializedName("rollbackFrom")
     * @Assert\Type("string")
     * @Type("string")
     */
    protected $rollbackFrom;

    /**
     * @var string|null
     * @SerializedName("rollbackTo")
     * @Assert\Type("string")
     * @Type("string")
     */
    protected $rollbackTo;

    /**
     * Constructor
     * @param mixed[] $data Associated array of property values initializing the model
     */
    public function __construct(array $data = null)
    {
        $this->target = isset($data['target']) ? $data['target'] : null;
        $this->rollbackFrom = isset($data['rollbackFrom']) ? $data['rollbackFrom'] : null;
        $this->rollbackTo = isset($data['rollbackTo']) ? $data['rollbackTo'] : null;
    }

    /**
     * Gets target.
     *
     * @return string
     */
    public function getTarget()
    {
        return $this->target;
    }

    /**
     * Sets target.
     *
     * @param string $target
     *
     * @return $this
     */
    public function setTarget($target)
    {
        $this->target = $target;

        return $this;
    }

    /**
     * Gets rollbackFrom.
     *
     * @return string|null
     */
    public function getRollbackFrom()
    {
        return $this->rollbackFrom;
    }

    /**
     * Sets rollbackFrom.
     *
     * @param string|null $rollbackFrom
     *
     * @return $this
     */
    public function setRollbackFrom($rollbackFrom = null)
    {
        $this->rollbackFrom = $rollbackFrom;

        return $this;
    }

    /**
     * Gets rollbackTo.
     *
     * @return string|null
     */
    public function getRollbackTo()
    {
        return $this->rollbackTo;
    }

    /**
     * Sets rollbackTo.
     *
     * @param string|null $rollbackTo
     *
     * @return $this
     */
    public function setRollbackTo($rollbackTo = null)
    {
        $this->rollbackTo = $rollbackTo;

        return $this;
    }
}